removed and changed some html. put css from homelab.php into own .css file
