+++ index.html
@@ -7,16 +7,16 @@
     <link rel="shortcut icon" type="image/ico" href="favicon.ico" />
     <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.6/css/bootstrap.min.css">
     <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.6/css/bootstrap-theme.min.css">
-    <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/font-awesome/4.5.0/css/font-awesome.min.css" /> <!--FontAwesome-->
-    <link href='//fonts.googleapis.com/css?family=PT+Sans:400' rel='stylesheet' type='text/css'> <!-- Font -->
-    <link rel="stylesheet" href="css/reset.css"> <!-- CSS reset -->
-    <link rel="stylesheet" href="css/style.css"> <!-- Resource style -->
+    <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/font-awesome/4.5.0/css/font-awesome.min.css" />
+    <link href='//fonts.googleapis.com/css?family=PT+Sans:400' rel='stylesheet' type='text/css'>
+    <link rel="stylesheet" href="css/reset.css">
+    <link rel="stylesheet" href="css/style.css">
   
-    <script src="js/modernizr.js"></script> <!-- Modernizr -->
+    <script src="js/modernizr.js"></script>
     <script src="js/jquery.min.js"></script>
     <script src="js/bootstrap.min.js"></script> -->
 
-    <title>Home lab</title>
+    <title><title><?php echo "($totalon/$totalcount) "; ?> Home lab</title></title>
 </head>
   
 <body>
@@ -87,9 +87,7 @@
 	</ul>      
                
 </div>
-<head><title><?php echo "($totalon/$totalcount) "; ?>Home lab</title></head>
-<script src="js/jquery-2.1.1.js"></script>
-<script src="js/main.js"></script> <!-- Resource jQuery -->
 
+<script src="js/main.js"></script>
 </body>
 </html>
